apps/backend/pyproject.toml
[tool.poetry]
name = "resonant"
version = "0.1.0"
description = ""
authors = ["Henry Nguyen <henry2833@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^1.10"
trafilatura = "^1.6.1"
aiohttp = "^3.8.5"
lmdb = "^1.4.1"
newspaper3k = "^0.2.8"
pytest = "^7.4.0"
autopep8 = "^2.0.2"
argparse = "^1.4.0"
nltk = "^3.8.1"
numpy = "^1.25.2"
validators = "^0.21.2"
beautifulsoup4 = "^4.12.2"
load-dotenv = "^0.1.0"
prisma = "^0.9.1"
sentence-transformers = "^2.2.2"
psycopg = {extras = ["binary"], version = "^3.1.10"}
gensim = "^4.3.2"
mmh3 = "^4.0.1"
pytest-asyncio = "^0.21.1"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
feedparser = "^6.0.10"
gunicorn = "^21.2.0"
langchain = "^0.0.327"
anthropic = "^0.5.0"
networkx = "^3.2.1"
plotly = "^5.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

